requirements.txt
'''fastapi
uvicorn
requests
chromadb
sentence-transformers
ollama
python-multipart
'''



fastapi
uvicorn
requests
python-multipart

chromadb
pysqlite3-binary

sentence-transformers
torch

ollama
python-dotenv
